+++ src/main/java/org/library/domain/Message.java
@@ -1,4 +1,4 @@
-package org.library.entity;
+package org.library.domain;
 
 public enum Message {
