+++ src/main/java/org/library/exception/InvalidModeException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class InvalidModeException extends RuntimeException {
 
     public InvalidModeException() {
