+++ src/test/java/org/library/ReportLostTest.java
@@ -1,16 +1,17 @@
 package org.library;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.library.domain.Message;
 import org.library.entity.Book;
-import org.library.entity.Message;
 import org.library.exception.AlreadyLostException;
 import org.library.repository.ApplicationRepository;
 import org.library.repository.Repository;
 import org.library.service.BookService;
-
-import static org.assertj.core.api.Assertions.*;
 
 public class ReportLostTest {
 
@@ -19,14 +20,14 @@ public class ReportLostTest {
     private String path = "src/test/resources/Book.json";
 
     @BeforeEach
-    void init(){
+    void init() {
         repository = new ApplicationRepository(path);
         service = new BookService(repository);
     }
 
     @DisplayName("도서 분실 처리")
     @Test
-    void 도서_분실_처리(){
+    void 도서_분실_처리() {
         //given
         Book book = new Book(service.generateId(), "testTitle", "testAuthor", 33);
 
@@ -40,7 +41,7 @@ public class ReportLostTest {
 
     @DisplayName("분실 상태에서 분실 처리")
     @Test
-    void 분실상태_분실처리(){
+    void 분실상태_분실처리() {
         //given
         Book book = new Book(service.generateId(), "testTitle", "testAuthor", 33);
