+++ src/test/java/org/library/ReturnsTest.java
@@ -1,16 +1,17 @@
 package org.library;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.library.domain.Message;
 import org.library.entity.Book;
-import org.library.entity.Message;
 import org.library.exception.NotReturnsException;
 import org.library.repository.ApplicationRepository;
 import org.library.repository.Repository;
 import org.library.service.BookService;
-
-import static org.assertj.core.api.Assertions.*;
 
 public class ReturnsTest {
 
@@ -19,14 +20,14 @@ public class ReturnsTest {
     private String path = "src/test/resources/Book.json";
 
     @BeforeEach
-    void init(){
+    void init() {
         repository = new ApplicationRepository(path);
         service = new BookService(repository);
     }
 
     @DisplayName("대여 상태에서 반납")
     @Test
-    void 대여상태_반납(){
+    void 대여상태_반납() {
         //given
         Book book = new Book(service.generateId(), "테스트이름", "테스트저자", 20);
 
@@ -40,7 +41,7 @@ public class ReturnsTest {
 
     @DisplayName("분실 상태에서 반납")
     @Test
-    void 분실상태_반납(){
+    void 분실상태_반납() {
         //given
         Book book = new Book(service.generateId(), "테스트이름", "테스트저자", 20);
 
@@ -54,7 +55,7 @@ public class ReturnsTest {
 
     @DisplayName("대여가능 상태에서 반납")
     @Test
-    void 대여가능상태_반납(){
+    void 대여가능상태_반납() {
         //given
         Book book = new Book(service.generateId(), "테스트이름", "테스트저자", 20);
 
@@ -62,12 +63,12 @@ public class ReturnsTest {
         service.save(book);
 
         //then
-        assertThatThrownBy(()->book.returns()).isInstanceOf(NotReturnsException.class);
+        assertThatThrownBy(() -> book.returns()).isInstanceOf(NotReturnsException.class);
     }
 
     @DisplayName("도서정리중 상태에서 반납")
     @Test
-    void 도서정리중상태_반납(){
+    void 도서정리중상태_반납() {
         //given
         Book book = new Book(service.generateId(), "테스트이름", "테스트저자", 20);
 
@@ -77,6 +78,6 @@ public class ReturnsTest {
         service.returns(book.getId());
 
         //then
-        assertThatThrownBy(()-> book.returns()).isInstanceOf(NotReturnsException.class);
+        assertThatThrownBy(() -> book.returns()).isInstanceOf(NotReturnsException.class);
     }
 }
